+++ android/app/src/main/java/com/alphalog/app/MainActivity.java
@@ -1,0 +1,5 @@
+package com.alphalog.app;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}